requirements.txt
absl-py==2.1.0
affine==2.4.0
alabaster==1.0.0
appdirs==1.4.4
astunparse==1.6.3
attrs==24.2.0
autograd==1.8.0
autoray==0.7.2
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cirq-core==1.3.0
cirq-google==1.3.0
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
commonmark==0.9.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.3.4
docutils==0.21.2
duet==0.2.9
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.22.0
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h5py==3.12.1
ibm-cloud-sdk-core==3.21.0
ibm-platform-services==0.57.1
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multimethod==1.12
multiprocess==0.70.12.2
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
pathos==0.2.8
pbr==6.1.0
pennylane==0.42.1
pennylane_lightning==0.42.0
pillow==10.4.0
platformdirs==4.3.6
ply==3.11
pox==0.3.5
ppft==1.7.6.9
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pylatexenc==2.10
pyparsing==3.1.4
pyscf==2.7.0
pyspnego==0.11.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qiskit==0.45.3
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.20.0
qiskit-terra==0.45.3
rasterio==1.4.3
recommonmark==0.7.1
referencing==0.35.1
requests==2.32.3
requests_ntlm==1.3.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
rustworkx==0.15.1
scikit-learn==1.6.1
scipy==1.14.1
scipy-openblas32==0.3.30.0.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stevedore==5.3.0
symengine==0.9.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-quantum==0.7.3
tensorpack==0.11
termcolor==2.5.0
threadpoolctl==3.6.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.3
torch==2.4.1
torchdiffeq==0.2.4
torchpack==0.3.1
torchquantum==0.1.8
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
